Added Rectangle to geometry package to minimally describe rectangles that may be drawn to screen.
Added Lawn to models package for drawing the rectangular lawn that features in the game.

Modified ModelBuilder with appendRectangle that specifies the OpenGL primitives and drawing instructions needed to composed a rectangle.

Modified GameRenderer, adding constants that maintain the sizes and proportions of various game models in terms of a 7 x 11 region made from square cells.  Adjusted draw frame to draw a lawn with one segment partially obscured by a patch of grass.  Also enabled alpha blending for necessary transparency effects.

Modified ColorShader to accept and use the alpha channel for colors so transparency effects can be applied to models drawn with it.

Added JUnit as an implementation dependency so an important assertion could be placed in GameRenderer that checks for cell width and height to match as they should.