Added SquareWithHole to minimally describe a square with a circle hole cut out of it.
Added GrassHole to models package for drawing holes in the grass on the game's lawn.

Corrected Lawn to append a rectangle instead of a square.  Type signatures got goofed up somehow, though it still worked.

Modified ModelBuilder with appendSquareWithHole.  It presently draws a quarter of a complete square with hole.

Modified GameRenderer to draw a GrassHole in the scene.