Modified FragmentGame with additional TextViews for Display Name and labeling Time.
Added values to Strings.

Modified GameFragment to capture additional GameMenu items related to signing in and out, and added functionality for Google Sign In with observers on the underlying ViewModel to update the Profile Pic and Display Name when signed in or out.

Created AccountViewModel with to maintain Google Sign-In data through lifecycle changes as live data that fragments can subscribe to.