Modified Point with functionality to detect if it intersects the path between two other points.
Added Vector to express direction and magnitude that can be applied to Points.

Added Centipede that tracks the Position, Speed, Direction, and other characteristics of a Centipede on a Lawn.  Centipede's can be above or below ground and have other Centipede's as a Head or Tail.

Modified GameRenderer to loop the Game ViewModel with a time slice from TimeUtil and then position Segements into the scene based on the Centipedes in it.

Added TimeUtil to track slices of elapsed time and convert between different time values.

Modifed Game ViewModel substantially.  Capitalized the constants.  Added positioned Centipedes to test animation, as well as mutable live data for scoring information.  Added the game loop which presently animates Centipedes when the game is Running, changing Centipede's to above or below ground when they encounter holes.