Added textures for better centipede appearance than mere solid colors can provide.
Constrained the GLSurfaceView's aspect ration in the layouts so the renderer can just fill the whole region and be alleviated of need for stencil testing to prevent drawing centipede segments not yet in the lawn when they enter it.

Added GLSL texture fragment and vertex shader programs.

Modified GameFragment to configure OpenGL with color and depth buffers at the device's native maximums rather than specifying bits.

Added appendTextureCircle to ModelBuilder which more or less emulates appendCircle but considers a different vertex count to allow for texture coordinates at each vertex.

Modified Segment to draw a textured circle instead of a colored circle so centipede's can have better screen appearance.

Modified GameRenderer, adding a texture shader program and textures for centipede head and body segments.  Changed how aspect is calculated to always fill the entire viewport and removed scaling no longer needed.  Migrated calls to colorShader.use to the positionXInScene methods that need it because positionSegmentInScene now uses textureShader.  Also modified positionModelinScene to accept a rotation about Z which is needed for centipede segments because their texture should be oriented in their direction of travel.