Redid centipede segment textures with circles and transparent surrounding.  Also renamed textures for better uniformity to make configuration through settings easier later.
Removed ColorModelBuilder as it is no longer used.  It's on GitHub in previous commits if needed later.

Changed Segment to build a rectangle versus a circle since the texture can accommodate the shape.  This improves performance.

Renamed a few things in GameRenderer.

Modified some constants in Game ViewModel to accommodate centipede's square shape.