Renamed some icons to better match their purpose.
Completed phone portrait and landscape layouts for the game over screen.

Added a BackFragment interface to GameActivity that fragments can optionally implement to be notified of back presses.  Implementing fragments can act on these presses and signal whether or not they have consumed it.  GameActivity checks the visible fragment for implementation and if so, notifies it and checks for consumption, calling the superclass implementation accordingly.

Fleshed out GameOverFragment to setup observers on the Game and Account ViewModels, updating the views according to the information available in them.  It also implements BackFragment so it can reset the game on back presses versus in onStop which would not work correctly.