Added additional textures for various centipede themes.
Added FragmentGameOver layouts for tablets.  These need additional work.

Modified Vector to provide a target rotation amount for particular directions.

Modified Centipede to track a next direction at its approach to a turn that gets written over its direction when it actually reaches it.  Also added fields to track its target rotation, rotation percentage, and actual rotation which is a the percentage applied to the target rotation.

Modified GameRenderer, removing rotation calculations for segments that belonged in Game ViewModel.

Modified Game ViewModel heavily.  It now figures out the rotation that a Centipede needs based on its approach to a turn and applies a percentage of it to its actual rotation as it moves through a turn.  This causes the centipedes to rotate smoothly.

Refactoring needs to be done to find efficiencies in the code with smooth rotation now working.