Modified GameFragment to show View Leaderboard menu item only when player is signed in with Google, as a signed in account is necessary to view this.
Corrected comments in GameOverFragment.

Fleshed out basic leaderboard retrieval functionality in LeaderboardFragment.  It presently retrieves and show the leaderboard entry for the signed in player, including top score, rounds, time, and placement, along with Google Play Games display name and avatar.  A message overlay gives indication when leaderboard entries are being retrieved and if this fails.

Added an icon for leaderboard retrieval.

Added ConstraintLeaderboard layout to show the the statistics of the signed in player's leaderboard entry.

Modified FragmentLeaderboard with message overlays and ConstraintLeaderboard.

Added an additional dimension for the signed in player's leaderboard entry avatar.

Added string resources for leaderboard entry retrieval status messages and statistics display.