Removed achievement images as they can be loaded as needed through facilities of the Play Games achievements client.
Added constraint_achievement layout for the display of achievements in a recyclerview.

Modified the layout of fragment_leaderboard for slightly better appearance at design and runtime.

Added additional strings and dimensions for achievement avatar images.

Modified LeaderboardFragment to setup the achievements recyclerview with achievements for the signed in player via PlayGamesUtil.

Added AchievementAdapter to adapter achievements in an achievement buffer for display, similar to LeaderboardAdapter.

Modified PlayGamesUtil with a method for retrieving achievements for a signed in player from Play Games.