Modified constraint layouts for GameOverFragment to remove old and unneeded constraints as they are contained in a frame layout.
Added several constraints for the various retrieval messages for the LeaderboardFragment and modified fragment_leaderboard to include them for more readability and succinctness.

Modified LeaderboardFragment to hide the player leaderboard entry when it is unavailable, which showed after changing the layout semantics of the screen.

Added a method to LeaderboardAdapter to indicate whether or not the signed in player was the only leader in the collection whom would not show because the player's entry is already shown at the top.

Modified fragment_leaderboard to contain everything in a scroll view for better display of all elements, especially in landscape orientation.