Added any_power_up.png to test all power ups with some common texture for proper display.
Added PowerUp which tracks the position, speed, and direction of a power up ala Centipede.

Added PowerUpRendering so GameRenderer knows how to draw a power up.

Modified the commentary in Segment, which was found not to match its functionality.

Modified GameRenderer which logic to position power ups in the scene based on those found in the Game ViewModel.

Added secondsToMillis to TimeUtil.

Modified GameViewModel.  Moved Random to a private static shared among routines that need random number generation to avoid needless GC.  Modified StartingEdge with its own functionality to return random edges, points along that edge, and directions into it.  Added tracking for Power Ups, along with methods to setup, kill, and apply them in much the same way centipedes are.  Modifed setupCentipede to use StartEdge's random methods.