Added constraint_theme_preview to be used by ThemePreviewPreference to show a preview of the selected theme.
Added additional key in strings for color theme preview preferences.

Modified game_settings to include a ThemePreviewPreference.

Added comment block to CreditsFragment and SettingsFragment.

Added ThemePreviewPreference to show previews of user configured themes.

Modified ThemeUtil to provide a configured theme based on the passed in name.  This is a work around the apparent race condition where user shared defaults do not contain the new theme when the theme's onPreferenceChanged handler is called.