Added HTML and PNG assets for instructions.
Modified GameFragment with navigation actions for the instructions menu item.

Added the InstructionsFragment which loads the FragmentInstructions' web view with the embedded HTML help.

Added IconInstructions for the GameMenu.

Added FragmentInstructions for displaying the game's HTML help.

Modified MenuGame with a View Instructions action.

Modified the navigation graph with a path from the game to the instructions fragment.